atualizando AuthorizationService

+++ src/main/java/com/example/weather/service/PoluicaoService.java
@@ -1,13 +1,12 @@
 package com.example.weather.service;
 
-import com.example.weather.converter.PoluicaoConverter;
-import com.example.weather.converter.PoluicaoMapper;
-import com.example.weather.entity.CidadeEntity;
 import com.example.weather.entity.PoluicaoEntity;
 import com.example.weather.reponse.PoluicaoResponseDTO;
 import com.example.weather.repository.PoluicaoRepository;
 import com.example.weather.request.PoluicaoRequestDTO;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,23 +19,18 @@ import static org.springframework.util.Assert.notNull;
 @RequiredArgsConstructor
 public class PoluicaoService {
 
-    private final PoluicaoRepository poluicaoRepository;
-    private final PoluicaoConverter poluicaoConverter;
-    private final PoluicaoMapper poluicaoMapper;
-
-
-    public PoluicaoEntity salvaPoluicao(PoluicaoEntity poluicaoEntity) {
-        return poluicaoRepository.save(poluicaoEntity);
-    }
+    @Autowired
+    private PoluicaoRepository poluicaoRepository;
 
     public PoluicaoResponseDTO gravarPoluicao(PoluicaoRequestDTO poluicaoRequestDTO) {
         try {
             notNull(poluicaoRequestDTO, "Os dados de poluicao são obrigatórios");
-            PoluicaoEntity poluicaoEntity = salvaPoluicao(poluicaoConverter.paraPoluicaoEntity(poluicaoRequestDTO));
+            PoluicaoEntity novaPoluicao = new PoluicaoEntity();
+            BeanUtils.copyProperties(poluicaoRequestDTO, novaPoluicao);
 
-            return poluicaoMapper.paraPoluicaoResponseDTO(poluicaoEntity);
+            PoluicaoEntity poluicaoSalvo = poluicaoRepository.save(novaPoluicao);
+            return new PoluicaoResponseDTO(poluicaoSalvo);
         } catch (Exception e) {
-            e.getMessage();
             throw new NullPointerException("Erro ao gravar dados da poluição");
         }
     }
@@ -45,26 +39,35 @@ public class PoluicaoService {
     public PoluicaoResponseDTO buscaDadosPoluicao(String id) {
         try {
             Optional<PoluicaoEntity> entity = poluicaoRepository.findById(id);
-            notNull(entity, "Poluição não encontrado");
 
-            return poluicaoMapper.paraPoluicaoResponseDTO(entity.get());
+            if (entity.isPresent()){
+                return new PoluicaoResponseDTO(entity.get());
+            } else {
+                throw new RuntimeException("Dados da poluição não existe!");
+            }
+
         } catch (Exception e) {
-            e.getMessage();
             throw new NullPointerException("Erro ao buscar dados de Poluição");
         }
     }
 
     public List<PoluicaoResponseDTO> buscarTodasPoluicoes(){
-        List<PoluicaoEntity> poluicoes = poluicaoRepository.findAll();
-
-        return poluicoes.stream().map(poluicaoMapper::paraPoluicaoResponseDTO).toList();
+        return poluicaoRepository
+                .findAll()
+                .stream()
+                .map(PoluicaoResponseDTO::new)
+                .toList();
 
     }
 
     @Transactional
     public void deletaDadosPoluicao(String id) {
         Optional<PoluicaoEntity> entity = poluicaoRepository.findById(id);
-        poluicaoRepository.deleteById(id);
+        if (entity.isPresent()){
+            poluicaoRepository.deleteById(id);
+        }  else {
+            throw new RuntimeException("Dados da poluição não existe!");
+        }
 
     }
